requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0